pyproject.toml
[project]
name = "trascrizioni"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faster-whisper>=1.2.1",
    "moviepy>=2.2.1",
    "ollama>=0.6.1",
    "python-docx>=1.2.0",
]
